+++ src/cascadia/TerminalApp/App.h
@@ -5,9 +5,9 @@
 
 #include "Tab.h"
 #include "CascadiaSettings.h"
+#include "TerminalPage.h"
 #include "App.g.h"
 #include "App.base.h"
-#include "ScopedResourceLoader.h"
 #include "../../cascadia/inc/cppwinrt_utils.h"
 
 #include <winrt/Microsoft.Terminal.TerminalControl.h>
@@ -25,8 +25,7 @@ namespace winrt::TerminalApp::implementation
     {
     public:
         App();
-
-        Windows::UI::Xaml::UIElement GetRoot() noexcept;
+        ~App() = default;
 
         void Create();
         void LoadSettings();
@@ -34,15 +33,15 @@ namespace winrt::TerminalApp::implementation
         Windows::Foundation::Point GetLaunchDimensions(uint32_t dpi);
         bool GetShowTabsInTitlebar();
 
-        ~App() = default;
+        Windows::UI::Xaml::UIElement GetRoot() noexcept;
 
-        hstring GetTitle();
+        hstring Title();
         void TitlebarClicked();
 
         // -------------------------------- WinRT Events ---------------------------------
-        DECLARE_EVENT(TitleChanged, _titleChangeHandlers, winrt::Microsoft::Terminal::TerminalControl::TitleChangedEventArgs);
-        DECLARE_EVENT(LastTabClosed, _lastTabClosedHandlers, winrt::TerminalApp::LastTabClosedEventArgs);
-        DECLARE_EVENT_WITH_TYPED_EVENT_HANDLER(SetTitleBarContent, _setTitleBarContentHandlers, TerminalApp::App, winrt::Windows::UI::Xaml::UIElement);
+        DECLARE_EVENT_WITH_TYPED_EVENT_HANDLER(TitleChanged, _titleChangeHandlers, winrt::Windows::Foundation::IInspectable, winrt::hstring);
+        DECLARE_EVENT_WITH_TYPED_EVENT_HANDLER(LastTabClosed, _lastTabClosedHandlers, winrt::Windows::Foundation::IInspectable, winrt::TerminalApp::LastTabClosedEventArgs);
+        DECLARE_EVENT_WITH_TYPED_EVENT_HANDLER(SetTitleBarContent, _setTitleBarContentHandlers, winrt::Windows::Foundation::IInspectable, winrt::Windows::UI::Xaml::UIElement);
         DECLARE_EVENT_WITH_TYPED_EVENT_HANDLER(RequestedThemeChanged, _requestedThemeChangedHandlers, TerminalApp::App, winrt::Windows::UI::Xaml::ElementTheme);
 
     private:
@@ -50,95 +49,36 @@ namespace winrt::TerminalApp::implementation
         // the ctor, you're going to have a bad time. It'll mysteriously fail to
         // activate the app.
         // ALSO: If you add any UIElements as roots here, make sure they're
-        // updated in _ApplyTheme. The two roots currently are _root and _tabRow
-        // (which is a root when the tabs are in the titlebar.)
-        Windows::UI::Xaml::Controls::Control _root{ nullptr };
-        Microsoft::UI::Xaml::Controls::TabView _tabView{ nullptr };
-        TerminalApp::TabRowControl _tabRow{ nullptr };
-        Windows::UI::Xaml::Controls::Grid _tabContent{ nullptr };
-        Windows::UI::Xaml::Controls::SplitButton _newTabButton{ nullptr };
+        // updated in _ApplyTheme. The root currently is _root.
+        winrt::com_ptr<TerminalPage> _root{ nullptr };
 
-        std::vector<std::shared_ptr<Tab>> _tabs;
+        std::shared_ptr<::TerminalApp::CascadiaSettings> _settings{ nullptr };
 
-        std::unique_ptr<::TerminalApp::CascadiaSettings> _settings;
+        std::shared_ptr<ScopedResourceLoader> _resourceLoader{ nullptr };
 
         HRESULT _settingsLoadedResult;
         winrt::hstring _settingsLoadExceptionText{};
 
         bool _loadedInitialSettings;
-        std::shared_mutex _dialogLock;
-
-        ScopedResourceLoader _resourceLoader;
 
         wil::unique_folder_change_reader_nothrow _reader;
 
+        std::shared_mutex _dialogLock;
+
         std::atomic<bool> _settingsReloadQueued{ false };
 
-        void _CreateNewTabFlyout();
-        void _OpenNewTabDropdown();
-
-        fire_and_forget _ShowDialog(const winrt::Windows::Foundation::IInspectable& titleElement,
-                                    const winrt::Windows::Foundation::IInspectable& contentElement,
-                                    const winrt::hstring& closeButtonText);
-        void _ShowOkDialog(const winrt::hstring& titleKey, const winrt::hstring& contentKey);
-        void _ShowAboutDialog();
+        fire_and_forget _ShowDialog(const winrt::Windows::Foundation::IInspectable& sender, winrt::Windows::UI::Xaml::Controls::ContentDialog dialog);
+        void _ShowLoadErrorsDialog(const winrt::hstring& titleKey, const winrt::hstring& contentKey, HRESULT settingsLoadedResult);
         void _ShowLoadWarningsDialog();
-        void _ShowLoadErrorsDialog(const winrt::hstring& titleKey, const winrt::hstring& contentKey);
+
+        void _OnLoaded(const IInspectable& sender, const Windows::UI::Xaml::RoutedEventArgs& eventArgs);
 
         [[nodiscard]] HRESULT _TryLoadSettings() noexcept;
         void _LoadSettings();
         void _OpenSettings();
-
-        void _HookupKeyBindings(TerminalApp::AppKeyBindings bindings) noexcept;
-
         void _RegisterSettingsChange();
         fire_and_forget _DispatchReloadSettings();
         void _ReloadSettings();
-
-        void _SettingsButtonOnClick(const IInspectable& sender, const Windows::UI::Xaml::RoutedEventArgs& eventArgs);
-        void _FeedbackButtonOnClick(const IInspectable& sender, const Windows::UI::Xaml::RoutedEventArgs& eventArgs);
-        void _AboutButtonOnClick(const IInspectable& sender, const Windows::UI::Xaml::RoutedEventArgs& eventArgs);
-
-        void _UpdateTabView();
-        void _UpdateTabIcon(std::shared_ptr<Tab> tab);
-        void _UpdateTitle(std::shared_ptr<Tab> tab);
-
-        void _RegisterTerminalEvents(Microsoft::Terminal::TerminalControl::TermControl term, std::shared_ptr<Tab> hostingTab);
-
-        void _CreateNewTabFromSettings(GUID profileGuid, winrt::Microsoft::Terminal::Settings::TerminalSettings settings);
-        winrt::Microsoft::Terminal::TerminalConnection::ITerminalConnection _CreateConnectionFromSettings(GUID profileGuid, winrt::Microsoft::Terminal::Settings::TerminalSettings settings);
-
-        void _OpenNewTab(std::optional<int> profileIndex);
-        void _DuplicateTabViewItem();
-        void _CloseFocusedTab();
-        void _CloseFocusedPane();
-        void _SelectNextTab(const bool bMoveRight);
-        bool _SelectTab(const int tabIndex);
-
-        void _SetFocusedTabIndex(int tabIndex);
-        int _GetFocusedTabIndex() const;
-
-        void _Scroll(int delta);
-        bool _CopyText(const bool trimTrailingWhitespace);
-        void _PasteText();
-        void _SplitVertical(const std::optional<GUID>& profileGuid);
-        void _SplitHorizontal(const std::optional<GUID>& profileGuid);
-        void _SplitPane(const Pane::SplitState splitType, const std::optional<GUID>& profileGuid);
-
-        // Todo: add more event implementations here
-        // MSFT:20641986: Add keybindings for New Window
-        void _ScrollPage(int delta);
-        void _ResizePane(const Direction& direction);
-        void _MoveFocus(const Direction& direction);
-
-        void _OnLoaded(const IInspectable& sender, const Windows::UI::Xaml::RoutedEventArgs& eventArgs);
-        void _OnTabSelectionChanged(const IInspectable& sender, const Windows::UI::Xaml::Controls::SelectionChangedEventArgs& eventArgs);
-        void _OnTabClosing(const IInspectable& sender, const Microsoft::UI::Xaml::Controls::TabViewTabClosingEventArgs& eventArgs);
-        void _OnTabItemsChanged(const IInspectable& sender, const Windows::Foundation::Collections::IVectorChangedEventArgs& eventArgs);
-        void _OnTabClick(const IInspectable& sender, const Windows::UI::Xaml::Input::PointerRoutedEventArgs& eventArgs);
-        void _OnContentSizeChanged(const IInspectable& sender, Windows::UI::Xaml::SizeChangedEventArgs const& e);
-
-        void _RemoveTabViewItem(const IInspectable& tabViewItem);
 
         void _ApplyTheme(const Windows::UI::Xaml::ElementTheme& newTheme);
 
@@ -150,30 +90,6 @@ namespace winrt::TerminalApp::implementation
         void _PasteFromClipboardHandler(const IInspectable& sender, const Microsoft::Terminal::TerminalControl::PasteFromClipboardEventArgs& eventArgs);
 
         static void _SetAcceleratorForMenuItem(Windows::UI::Xaml::Controls::MenuFlyoutItem& menuItem, const winrt::Microsoft::Terminal::Settings::KeyChord& keyChord);
-
-#pragma region ActionHandlers
-        // These are all defined in AppActionHandlers.cpp
-        void _HandleNewTab(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleOpenNewTabDropdown(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleDuplicateTab(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleCloseTab(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleClosePane(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleScrollUp(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleScrollDown(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleNextTab(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandlePrevTab(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleSplitVertical(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleSplitHorizontal(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleScrollUpPage(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleScrollDownPage(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleOpenSettings(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandlePasteText(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleNewTabWithProfile(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleSwitchToTab(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleResizePane(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleMoveFocus(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-        void _HandleCopyText(const IInspectable& sender, const TerminalApp::ActionEventArgs& args);
-#pragma endregion
     };
 }
