+++ src/cascadia/WindowsTerminal/AppHost.h
@@ -14,8 +14,8 @@ public:
     AppHost() noexcept;
     virtual ~AppHost();
 
-    void AppTitleChanged(winrt::hstring newTitle);
-    void LastTabClosed();
+    void AppTitleChanged(const winrt::Windows::Foundation::IInspectable& sender, winrt::hstring newTitle);
+    void LastTabClosed(const winrt::Windows::Foundation::IInspectable& sender, const winrt::TerminalApp::LastTabClosedEventArgs& args);
     void Initialize();
 
 private:
@@ -25,7 +25,7 @@ private:
     winrt::TerminalApp::App _app;
 
     void _HandleCreateWindow(const HWND hwnd, const RECT proposedRect);
-    void _UpdateTitleBarContent(const winrt::TerminalApp::App& sender,
+    void _UpdateTitleBarContent(const winrt::Windows::Foundation::IInspectable& sender,
                                 const winrt::Windows::UI::Xaml::UIElement& arg);
     void _UpdateTheme(const winrt::TerminalApp::App&,
                       const winrt::Windows::UI::Xaml::ElementTheme& arg);
